refactor: move screens into dedicated folder
