+++ src/screens/PainelPrincipal.tsx
@@ -10,7 +10,7 @@ import {
   Package,
   ArrowLeft,
 } from 'lucide-react';
-import { Modal } from './Modal';
+import { Modal } from '../components/Modal';
 import { useScrollToTop } from '../hooks/useScrollToTop';
 import { useState } from 'react';
 import { Espetinho } from '../types/Espetinho';
