+++ src/screens/VisualizarOperacao.tsx
